[TEMP] feat(Canvas): Add containers for branches
Currently, all children are rendered as individual nodes in the Canvas,
making complicated to understand when a step belongs to a parent EIP or
if it comes as a next step in the Camel Route.

This commit adds support for rendering EIP branches as containers.

fix: https://github.com/KaotoIO/kaoto/issues/368